% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDiff.R
\name{calcDiff}
\alias{calcDiff}
\title{Diffusion Rate Calculation}
\usage{
calcDiff(growpars, svec, prey, pred)
}
\arguments{
\item{growpars}{parameters for growth function (r,D)}

\item{svec}{vector of size classes (upper bounds)}

\item{prey}{vector of prey densities}

\item{pred}{vector of predator densities}
}
\description{
Diffusion Rate Calculation
}
\examples{
calcDiff()
}
