% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPred.R
\name{calcPred}
\alias{calcPred}
\title{Predation Rate Calculation}
\usage{
calcPred(parmat, svec, preyvec, predvec, afun = inda, hfun = indh)
}
\arguments{
\item{parmat}{matrix of parameters that define predator attack rates and handling times}

\item{svec}{vector of size classes (upper bounds)}

\item{preyvec}{vector of prey densities}

\item{predvec}{vector of predator densities}

\item{afun}{attack-rate function}

\item{hfun}{handling time function}
}
\description{
Predation Rate Calculation
}
\examples{
svec <- 1:10
preyvec <- rep(1,10)
predvec <- 1
parmat <- matrix(c(1,1,1,   ## c,d,g pred 1
                   1,1,1),  ## c,d,g pred2
       byrow=TRUE,
       nrow=2,dimnames=list(c("pred1","pred2"),c("c","d","g")))
calcPred()
}
