% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expona.R
\name{expona}
\alias{expona}
\title{Exponential Attack Rate Function}
\usage{
expona(s, c, d, g)
}
\arguments{
\item{s}{size vector of prey}

\item{c}{maximum attack rate}

\item{d}{size at maximum attack rate (scaling factor along size classes)}

\item{g}{scaler parameter}
}
\description{
Exponential Attack Rate Function
}
\examples{
## expona()
}
