% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponh.R
\name{exponh}
\alias{exponh}
\alias{indh}
\alias{powFun1}
\title{Handling time functions}
\usage{
exponh(s, m, n)

indh(s, m, n)

powFun1(s, m, n)
}
\arguments{
\item{s}{size vector of prey}

\item{m}{handling time}

\item{n}{scaler parameter}
}
\description{
Handling time functions
}
\examples{
exponh(1,1,1)
}
