% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FR_Data_Sim.R
\name{FR_Data_Sim}
\alias{FR_Data_Sim}
\title{Functional Response Experiment Simulation}
\usage{
FR_Data_Sim(
  afun,
  FRfun,
  a,
  a_sd,
  h,
  d,
  gamma,
  reps,
  N0s,
  Sizes,
  tval,
  deterministic = FALSE,
  logRE = TRUE
)
}
\arguments{
\item{afun}{Size-dependent attack rate function}

\item{FRfun}{Hollings or Rogers}

\item{a}{maximum attack rate}

\item{h}{handling time}

\item{d}{size at max attack rate}

\item{gamma}{shape parameter for size dependence}

\item{reps}{number of replicates (or blocks) in the experiment}

\item{N0s}{vector of initial densities}

\item{Sizes}{vector of size classes}

\item{tval}{number of timesteps in experiment}

\item{deterministic}{return probability instead of binomal RV}

\item{logRE}{implement random effects on log scale}

\item{a_var}{random effect variance}
}
\description{
Functional Response Experiment Simulation
}
\examples{
FR_Data_Sim()
}
