% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FR_Disp.R
\name{FR_Disp}
\alias{FR_Disp}
\title{Functional Dispersion - Functional Response Parameters}
\usage{
FR_Disp(parms)
}
\arguments{
\item{parms}{matrix of parameters (currently 4 parameters, will recode)}
}
\description{
Functional Dispersion - Functional Response Parameters
}
\examples{
FR_Disp(rbind(1:4))
}
