% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gompertz.R
\name{Gompertz}
\alias{Gompertz}
\title{Gompertz Growth Function}
\usage{
Gompertz(max_size, s_init, r, time)
}
\arguments{
\item{max_size}{maximum size of prey}

\item{s_init}{initial size vector of prey}

\item{r}{growth rate}

\item{time}{time steps}
}
\description{
Gompertz Growth Function
}
\examples{
Gompertz(max_size=100,s_init=1,r=1,time=0:10)
}
