% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gompertz_size.R
\name{Gompertz_size}
\alias{Gompertz_size}
\title{Prey Growth Function}
\usage{
Gompertz_size(max_size, s_init, r)
}
\arguments{
\item{max_size}{maximum size of prey}

\item{s_init}{initial size vector of prey}

\item{r}{growth rate}

\item{time}{time steps}
}
\description{
Prey Growth Function
}
\examples{
Gompertz_size(max_size=100,s_init=1,r=1)
}
