% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientfun.R
\name{gradientfun}
\alias{gradientfun}
\title{Reorganize Parameters}
\usage{
gradientfun(time, x, params, afun = inda, hfun = indh)
}
\arguments{
\item{time}{vector timesteps for integration (between each reproduction event)}

\item{x}{initial condition vector (predator densities, prey densities)}

\item{params}{parameter list}

\item{afun}{attack rate function - default is size-independent}

\item{hfun}{handling time function - default is size-independent}
}
\description{
Reorganize Parameters
}
\examples{
## gradientfun()
}
