% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperbol.R
\name{hyperbol}
\alias{hyperbol}
\title{Hyperbolic Attack Rate Function}
\usage{
hyperbol(s, c, d, g)
}
\arguments{
\item{s}{size vector of prey}

\item{c}{maximum attack rate}

\item{d}{size at maximum attack rate (scaling factor along size classes)}

\item{g}{scaler parameter}
}
\description{
Hyperbolic Attack Rate Function
}
\examples{
## hyperbol()
}
