% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inda.R
\name{inda}
\alias{inda}
\title{Independent attack rate functions}
\usage{
inda(s, c, d, g)
}
\arguments{
\item{s}{size vector of prey}

\item{c}{maximum rate}

\item{d}{size at maximum attack rate (scaling factor along size classes)}

\item{g}{scaler parameter (for power ricker)}
}
\description{
Independent attack rate functions
}
