% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indh.R
\name{indh}
\alias{indh}
\title{Size-independent handling time}
\usage{
indh(s, m, n)
}
\arguments{
\item{s}{size vector of prey}

\item{m}{handling time}

\item{n}{scaler parameter}
}
\description{
Size-independent handling time
}
\examples{
## indh()
}
