% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.R
\name{lin}
\alias{lin}
\title{Linear handling time}
\usage{
lin(s, m, n)
}
\arguments{
\item{s}{size vector of prey}

\item{m}{handling time}

\item{n}{scaler parameter}
}
\description{
Linear handling time
}
\examples{
## lin()
}
