% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ling.R
\name{ling}
\alias{ling}
\title{Linear somatic growth (constant growth rate among sizes)}
\usage{
ling(max_size, s_init, r, time)
}
\arguments{
\item{max_size}{maximum size of prey}

\item{s_init}{initial size of prey}

\item{r}{prey growth rate}

\item{time}{timestep for growth}
}
\description{
Linear somatic growth (constant growth rate among sizes)
}
\examples{
ling()
}
