% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modlog.R
\name{modlog}
\alias{modlog}
\title{Modified Logistic Function}
\usage{
modlog(s, c, d, g)
}
\arguments{
\item{s}{size vector of prey}

\item{c}{maximum attack rate}

\item{d}{size at maximum attack rate (scaling factor along size classes)}

\item{g}{scaler parameter}
}
\description{
Modified Logistic Function
}
\examples{
## modlog()
}
