% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramSkelfun.R
\name{paramSkelfun}
\alias{paramSkelfun}
\title{Reorganize Parameters}
\usage{
paramSkelfun(
  npred,
  predpars = c("c", "d", "g", "m", "n"),
  growpars = c("r", "max_size", "D")
)
}
\arguments{
\item{npred}{number of predators}

\item{predpars}{vector of predation parameters}

\item{growpars}{vector of growth parameters}
}
\description{
Reorganize Parameters
}
\examples{
## paramSkelfun()
}
