% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powRicker2.R
\name{powRicker2}
\alias{powRicker2}
\title{Alternative Formation of Power Ricker Attack Rate}
\usage{
powRicker2(s, c, d, g)
}
\arguments{
\item{s}{size vector of prey}

\item{c}{maximum attack rate}

\item{d}{size at maximum attack rate (scaling factor along size classes)}

\item{g}{scaler parameter}
}
\description{
Alternative Formation of Power Ricker Attack Rate
}
\examples{
## powRicker2()
}
