% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pred_Sims.R
\name{Pred_Sims}
\alias{Pred_Sims}
\title{Simulate predation by multiple predators with size-dependent functional responses}
\usage{
Pred_Sims(
  maxsize,
  dd1,
  r1,
  prey_dens,
  new_prey,
  svec0,
  pred_dens,
  predpar.mat,
  timevec,
  num_steps,
  afun = inda,
  hfun = indh,
  ...
)
}
\arguments{
\item{maxsize}{maximum size (all bins will approach this)}

\item{dd1}{diffusion rate}

\item{r1}{growth rate}

\item{prey_dens}{density of prey (one number)}

\item{new_prey}{number of new juveniles}

\item{svec0}{initial size vector}

\item{pred_dens}{vector of predator density}

\item{predpar.mat}{matrix of predator functional response parameters (each line is a predator, each column is a parameter)}

\item{timevec}{vector of timesteps for ode (time between reproduction events)}

\item{num_steps}{number of overall timesteps (or reproduction events)}

\item{afun=inda}{attack rate functional form (default is size-independent attack rate)}

\item{hfun=indh}{handling time functional form (default is size-independent handling time)}
}
\description{
Simulate predation by multiple predators with size-dependent functional responses
}
\examples{
Pred_Sims()
}
