% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ricker.R
\name{Ricker}
\alias{Ricker}
\title{Ricker attack rate functions}
\usage{
Ricker(s, c, d, g)
}
\arguments{
\item{s}{size vector of prey}

\item{c}{maximum rate}

\item{d}{size at maximum attack rate (scaling factor along size classes)}

\item{g}{scaler parameter (for power ricker)}
}
\description{
Ricker attack rate functions
}
