% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogers_preds.R
\name{rogers_preds}
\alias{rogers_preds}
\title{Rogers per-capital mortality risk with multiple of the same predator}
\usage{
rogers_preds(N0, a, h, P, tval)
}
\arguments{
\item{N0}{initial prey density}

\item{a}{attack rate}

\item{h}{handling time}

\item{P}{number of predators}

\item{tval}{length of predation experiment}

\item{s}{size of prey}

\item{d}{size at max attack rate}

\item{gamma}{shape parameter for power ricker attack rate}

\item{afun}{form of attack rate size-dependence:  options, Ricker, powRicker (will expand later)}
}
\description{
Rogers per-capital mortality risk with multiple of the same predator
}
\examples{
rogers_risk()
}
