% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogers_risk.R
\name{rogers_risk}
\alias{rogers_risk}
\title{Hollings per-capital mortality risk}
\usage{
rogers_risk(N0, c, h, tval, s, d, gamma, afun)
}
\arguments{
\item{N0}{initial prey density}

\item{h}{handling time}

\item{tval}{length of predation experiment}

\item{s}{size of prey}

\item{d}{size at max attack rate}

\item{gamma}{shape parameter for power ricker attack rate}

\item{afun}{form of attack rate size-dependence:  options, Ricker, powRicker (will expand later)}

\item{a}{attack rate}
}
\description{
Hollings per-capital mortality risk
}
\examples{
rogers_risk()
}
