% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_attack.R
\name{size_attack}
\alias{size_attack}
\title{Size-dependent attack rates}
\arguments{
\item{s}{size vector of prey}

\item{c}{maximum rate}

\item{d}{size at maximum attack rate (scaling factor along size classes)}

\item{g}{scaler parameter (for power ricker)}
}
\description{
returns size-dependent attack rate
}
