% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponh.R
\name{exponh}
\alias{exponh}
\title{Exponential handling time}
\usage{
exponh(s, m, n)
}
\arguments{
\item{s}{size vector of prey}

\item{m}{handling time}

\item{n}{scaler parameter}
}
\description{
Exponential handling time
}
\examples{
exponh(1,1,1)
}
