% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inda.R
\name{inda}
\alias{inda}
\title{Size-independent attack rate (constant)}
\usage{
inda(s, c, d, g)
}
\arguments{
\item{s}{size vector of prey}

\item{c}{maximum attack rate}

\item{d}{size at maximum attack rate (scaling factor along size classes)}

\item{g}{scaler parameter}
}
\description{
Size-independent attack rate (constant)
}
\examples{
inda()
}
