% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powFun1.R
\name{powFun1}
\alias{powFun1}
\title{Exponential Handling time functions}
\usage{
powFun1(s, m, n)
}
\arguments{
\item{s}{size vector of prey}

\item{m}{handling time}

\item{n}{scaler parameter}
}
\description{
Exponential Handling time functions
}
