% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powRicker.R
\name{powRicker}
\alias{powRicker}
\title{Power Ricker Attack Rate}
\usage{
powRicker(s, c, d, g)
}
\arguments{
\item{s}{size vector of prey}

\item{c}{maximum attack rate}

\item{d}{size at maximum attack rate (scaling factor along size classes)}

\item{g}{scaler parameter}
}
\description{
Power Ricker Attack Rate
}
\examples{
## powRicker()
}
