% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powRicker.R, R/size_funs.R
\name{powRicker}
\alias{powRicker}
\alias{Ricker}
\alias{expona}
\alias{powRicker2}
\alias{hyperbol}
\alias{inda}
\title{Attack rate functions}
\usage{
powRicker(s, c, d, g)

Ricker(s, c, d, g)

expona(s, c, d, g)

powRicker2(s, c, d, g)

hyperbol(s, c, d, g)

inda(s, c, d, g)

powRicker(s, c, d, g)
}
\arguments{
\item{s}{size vector of prey}

\item{c}{maximum attack rate}

\item{d}{size at maximum attack rate (scaling factor along size classes)}

\item{g}{scaler parameter}
}
\description{
Attack rate functions

Power Ricker Attack Rate
}
\examples{
powRicker()
}
