% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ricker.R
\name{Ricker}
\alias{Ricker}
\title{Ricker Attack Rate Function}
\usage{
Ricker(s, c, d, g)
}
\arguments{
\item{s}{size vector of prey}

\item{c}{maximum attack rate}

\item{d}{size at maximum attack rate (scaling factor along size classes)}

\item{g}{scaler parameter}
}
\description{
Ricker Attack Rate Function
}
\examples{
Ricker(1,1,1,1)
}
